Rename account oauth DomainContract to AccountOAuthDomainContract
