+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/AccountOAuthModule.kt
@@ -3,8 +3,8 @@ package app.k9mail.feature.account.oauth
 import app.k9mail.core.common.coreCommonModule
 import app.k9mail.feature.account.oauth.data.AuthorizationRepository
 import app.k9mail.feature.account.oauth.data.AuthorizationStateRepository
-import app.k9mail.feature.account.oauth.domain.DomainContract
-import app.k9mail.feature.account.oauth.domain.DomainContract.UseCase
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract.UseCase
 import app.k9mail.feature.account.oauth.domain.usecase.CheckIsGoogleSignIn
 import app.k9mail.feature.account.oauth.domain.usecase.FinishOAuthSignIn
 import app.k9mail.feature.account.oauth.domain.usecase.GetOAuthRequestIntent
@@ -25,13 +25,13 @@ val featureAccountOAuthModule: Module = module {
         )
     }
 
-    factory<DomainContract.AuthorizationRepository> {
+    factory<AccountOAuthDomainContract.AuthorizationRepository> {
         AuthorizationRepository(
             service = get(),
         )
     }
 
-    factory<DomainContract.AuthorizationStateRepository> {
+    factory<AccountOAuthDomainContract.AuthorizationStateRepository> {
         AuthorizationStateRepository()
     }
