+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/domain/usecase/FinishOAuthSignIn.kt
@@ -1,12 +1,12 @@
 package app.k9mail.feature.account.oauth.domain.usecase
 
 import android.content.Intent
-import app.k9mail.feature.account.oauth.domain.DomainContract
-import app.k9mail.feature.account.oauth.domain.DomainContract.UseCase
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract.UseCase
 import app.k9mail.feature.account.oauth.domain.entity.AuthorizationResult
 
 class FinishOAuthSignIn(
-    private val repository: DomainContract.AuthorizationRepository,
+    private val repository: AccountOAuthDomainContract.AuthorizationRepository,
 ) : UseCase.FinishOAuthSignIn {
     override suspend fun execute(intent: Intent): AuthorizationResult {
         val response = repository.getAuthorizationResponse(intent)
