+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/domain/usecase/GetOAuthRequestIntent.kt
@@ -1,12 +1,12 @@
 package app.k9mail.feature.account.oauth.domain.usecase
 
 import app.k9mail.core.common.oauth.OAuthConfigurationProvider
-import app.k9mail.feature.account.oauth.domain.DomainContract
-import app.k9mail.feature.account.oauth.domain.DomainContract.UseCase.GetOAuthRequestIntent
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract.UseCase.GetOAuthRequestIntent
 import app.k9mail.feature.account.oauth.domain.entity.AuthorizationIntentResult
 
 internal class GetOAuthRequestIntent(
-    private val repository: DomainContract.AuthorizationRepository,
+    private val repository: AccountOAuthDomainContract.AuthorizationRepository,
     private val configurationProvider: OAuthConfigurationProvider,
 ) : GetOAuthRequestIntent {
     override fun execute(hostname: String, emailAddress: String): AuthorizationIntentResult {
