+++ feature/account/oauth/src/main/kotlin/app/k9mail/feature/account/oauth/domain/usecase/SuggestServerName.kt
@@ -1,7 +1,7 @@
 package app.k9mail.feature.account.oauth.domain.usecase
 
 import app.k9mail.core.common.mail.Protocols
-import app.k9mail.feature.account.oauth.domain.DomainContract.UseCase
+import app.k9mail.feature.account.oauth.domain.AccountOAuthDomainContract.UseCase
 
 @Deprecated("This is not needed anymore, remove once auth setup flow is updated")
 class SuggestServerName : UseCase.SuggestServerName {
